other fixes
