+++ tests/settings.go
@@ -4,4 +4,4 @@ var Port = 7540
 var DBFile = "./database/scheduler.db"
 var FullNextDate = true
 var Search = true
-var Token = `eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJleHBpcmF0aW9uIjoxNzM1NjIzMTE1LCJwYXNzd29yZCI6IjEyMzQ1In0.YkLOEqd-ZAXUVbe99i7mitxijVtHtbL0e1onwjtBRYc`
+var Token = `eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJleHBpcmF0aW9uIjoxNzM1NjMzMzQ3LCJwYXNzd29yZCI6IjEyMzQ1In0.l-9uMoCOWPEzf88dTj3Jl1DJWdGJI1c7NivMKqQpWSU`
